move some functions to util files to clean up the store file
